Move all logic from main to a seperate class
